backend/requirements.txt
# Flask - Web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Python environment variables
python-dotenv==1.0.0

# Additional utilities
Werkzeug==3.0.1

# Optional: For production server
# Gunicorn==21.2.0
